Was playing around with a demo
Was building a demo for tower tank. Got mega upset because javascript was being an ass. Set it up to allow for a player to move around a play area that is larger than the screen and made it so there is a flow field that generates whenever the player moves to an adjacent grid square. Very fun project. Very annoying issue.